+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-
+* Fixed **format** to use git as a default value.
 # 1.3.8
 * Updated the **secrets** command to work on forked branches.
 * Added a validation verifying that the pack's README.md file is not equal to pack description.